Keep CLI PCV in separate file from graphical PCV
